% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MAF.R
\name{filter_MAF}
\alias{filter_MAF}
\title{filter MAF}
\usage{
filter_MAF(
  MAF,
  filterMC3 = F,
  MC3_name = "MC3_Overlap",
  addAltFreq = F,
  t_depth = "t_depth",
  t_alt_count = "t_alt_count",
  maxNum = 3000,
  minNum = 0,
  patientsID = NULL,
  convert_coordinate = F,
  chain = "hg38ToHg19.over.chain"
)
}
\arguments{
\item{MAF}{MAF file, with relevant fields consisten with the GDC MAF specification (\url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/}).}

\item{filterMC3}{whether filtering mutation not in MC3, default is FALSE. see details.}

\item{MC3_name}{The colnames of MC3. If the MC3 is TRUE, this mutation is retained.}

\item{addAltFreq}{whether add the alt frequency of mutations, default is FALSE.}

\item{t_depth}{The colnames of t_depth.}

\item{t_alt_count}{The colnames of t_alt_count.}

\item{maxNum}{patients with mutations more than this number will be removed.}

\item{minNum}{patients with mutations less than this number will be removed.}

\item{patientsID}{removing patients without in patients ID.}

\item{convert_coordinate}{whether convert hg38 to 19(or hg19 to hg38), default is FALSE.}

\item{chain}{chain file "hg38ToHg19.over.chain"}
}
\value{
MAF after filtering.
}
\description{
Input is MAF, output is MAF.
}
\details{
we convert data from hg38 to hg19 , keep only SNP data , removing hyper-mutated patients.

the functions are including: (1)hg_converter (2)unique_tumor_addition_function (3)tumor_allele_adder (4)DNP_TNP_remover (5)removing_patients (6)getAltFreq

\strong{MC3:The Multi-Center Mutation Calling in Multiple Cancers}

The MC3 data set provides consistent variant calling and filtering across the 10K patients in The Cancer Genome Atlas (TCGA). see \url{https://gdc.cancer.gov/about-data/publications/mc3-2017}
}
\examples{
maf <- read.delim(
 file = system.file("extdata", "TCGA.KIRC.merge.combine.uniq.amaf", package = "MafData") ,
 header = T,skip = 1,stringsAsFactors = F)
chain = system.file("extdata", "hg38ToHg19.over.chain", package = "MafData")
maf_left = filter_MAF(MAF = maf, hg38to19 = T, chain = chain )
}
