% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_RefCDS_sum.R
\name{Generate_RefCDS_sum}
\alias{Generate_RefCDS_sum}
\title{Generate_RefCDS_sum}
\usage{
Generate_RefCDS_sum(RefCDS = RefCDS, out = out)
}
\arguments{
\item{RefCDS}{RefCDS}

\item{out}{saved file names}
}
\description{
Generate the RefCDS for the Ki/Ks model.
}
\details{
load data before running this code.
}
