% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAltFreq.R
\name{getAltFreq}
\alias{getAltFreq}
\title{getAltFreq}
\usage{
getAltFreq(MAF, t_depth = "t_depth", t_alt_count = "t_alt_count")
}
\arguments{
\item{MAF}{MAF file, with relevant fields consisten with the GDC MAF specification (\url{https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/}).}

\item{t_depth}{The colnames of t_depth.}

\item{t_alt_count}{The colnames of t_alt_count.}
}
\value{
MAF after filtering.
}
\description{
calculating the alternative allele frequency of mutations.
}
