% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObsEI.R
\name{getObsEI}
\alias{getObsEI}
\title{getObsEI}
\usage{
getObsEI(list = list(), CancerName = NULL)
}
\arguments{
\item{list}{lists of KiKs_cancers}

\item{CancerName}{The corresponding cancer names}
}
\value{
lists including obsEI and obsGlobal
}
\description{
summary a lot of KiKs_cancer resluts into a list.
}
