% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaKs_patients.R
\name{KaKs_patients}
\alias{KaKs_patients}
\title{KaKs_patients}
\usage{
KaKs_patients(dndscvOUT, patientsID = NULL, geneID = NULL, sm = "192r_3w")
}
\arguments{
\item{dndscvOUT}{output from the dndscv, see \code{\link[dndscv]{dndscv}}}

\item{patientsID}{the patients used for the Ka/Ks calculation.}

\item{geneID}{the genes used for the Ka/Ks calculation.}

\item{sm}{Please set the sm same with pre-calculation.}
}
\value{
KaKs_patients returns a list of objects:

- PatientsKaKs: patients Ka/Ks.

- globaldnds: globaldnds information.
}
\description{
calculate the Ka/Ks per cancer case. The default is to use all genes across all patients to calculate the Ka/Ks per cancer case.
}
\details{
The first step is calculate the calculate \code{\link[dndscv]{dndscv}} from the whole dataset, and then extract the information from the dndscv output.

The Ka/Ks per patients is equal to the number of A/S changes per the site of #A/#S. The A/S changes is from each patient and the #A/#S site is designated for the genome-wide #A/#S site.

The Ka/Ks can be calculated from a specific gene list rather a genome-wide Ka/Ks, thus we can set the geneID for the specific gene list.

In the other situation, the Ka/Ks is based on a certain patients rather than whole patients. This will affect the genome-wide #A/#S site. Set the patientsID for a certain group of patients.
}
