% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KiKs_cancer.R
\name{KiKs_cancer}
\alias{KiKs_cancer}
\title{KiKs_cancer}
\usage{
KiKs_cancer(
  mutations = NULL,
  maf = NULL,
  hg38to19 = F,
  profile_mle = NULL,
  gene_list = NULL,
  syntofew = F,
  refdb = "hg19",
  sm = "192r_3w",
  kc = "cgc81",
  cv = "hg19",
  max_muts_per_gene_per_sample = 10,
  max_coding_muts_per_sample = 3000,
  use_indel_sites = T,
  min_indels = 5,
  maxcovs = 20,
  constrain_wnon_wspl = T,
  outp = 1,
  KiKsmodel = T
)
}
\arguments{
\item{mutations}{Environment input, Table of mutations (5 columns: sampleID, chr, pos, ref, alt). Only list independent events as mutations.}

\item{maf}{file names, with columns: "chr","pos","ref","mut","sampleID", separated by "Tab", without header.}

\item{hg38to19}{F; convert hg38 to 19; In default, mutations are based on hg19;}

\item{profile_mle}{profile estimated by mle of possion regression, the output \strong{mutrates} could as input of this params.}

\item{gene_list}{List of genes to restrict the analysis (use for targeted sequencing studies)}

\item{syntofew}{F (default). In case of too few mutations in Syn, thus we use the Ki1 to replace the SYN. This will only work when your data are filtered out of SYN mutations.}

\item{refdb}{Reference database (path to .rda file)}

\item{sm}{Substitution model (precomputed models are available in the data directory); sm = c("Type1_7_CpG","Type2_13_CpG","Type3_17","Type4_96","Type5_192"). see \strong{details} also.}

\item{kc}{List of a-priori known cancer genes (to be excluded from the indel background model)}

\item{cv}{Covariates (a matrix of covariates -columns- for each gene -rows-) [default: reference covariates] [cv=NULL runs dndscv without covariates]}

\item{max_muts_per_gene_per_sample}{If n<Inf, arbitrarily the first n mutations by chr position will be kept}

\item{max_coding_muts_per_sample}{Hypermutator samples often reduce power to detect selection}

\item{use_indel_sites}{Use unique indel sites instead of the total number of indels (it tends to be more robust)}

\item{min_indels}{Minimum number of indels required to run the indel recurrence module}

\item{maxcovs}{Maximum number of covariates that will be considered (additional columns in the matrix of covariates will be excluded)}

\item{constrain_wnon_wspl}{This constrains wnon==wspl (this typically leads to higher power to detect selection)}

\item{outp}{1 default; Output: 1 = Global dN/dS values; 2 = Global dN/dS and dNdSloc, 3 =  Global dN/dS, dNdSloc and dNdScv, 4 = 3 + socre test . see \strong{details}.}

\item{KiKsmodel}{run Ki/Ks (default = T);}
}
\value{
'dndscv' returns a list of objects:

Ki_mle: a list of variables. both likelihood ratio test and score test are applied.

globaldnds: globaldnds

sel_cv: dNdSloc

mutations: annotated mutations

p1: plot KI against UI.

UI.cor: correlation between 75-KI and UI.

mutrates:  mutation rate matrix, could as input of \strong{profile_mle}

mle_submodel: the estimated par from mle model of possion mutation rate regression

poissmodel: mle model of possion mutation rate regression

genemuts: mutations summarized by genes

par_KI: KI test by including w1-75 into one poisson regression model.

sel_loc: dNdScv

nbreg: negative binomial model for dNdScv
}
\description{
Analyses of selection using the dNdScv and dNdSloc models. Default parameters typically increase the performance of the method on cancer genomic studies. Reference files are currently only available for the GRCh37/hg19 version of the human genome.\cr
This is the function based on \code{\link[dndscv]{dndscv}} package. Here, we calculate the Ki/Ks(i=1:75) based on poisson regression model, similar with dNdSloc model in Amino acid levels. Except likelihood ratio test for selection, we also include a score test. see \strong{Details}.
}
\details{
load data before running this code.

data("RefCDS_hg19_sum", package="MafData.Data")

data("AminoAcidDistance", package="MafData.Data")

\strong{score test} score test is more strict than likelihood test(LR).

\strong{sm}, Substitution model: 1) sm =1:5; sm = c("Type1_7_CpG","Type2_13_CpG","Type3_17","Type4_96","Type5_192") or 2) sm ="192r_3w", "12r_3w","2r_3w" or created by yourselves.
}
\references{
Martincorena I, et al. (2017) Universal Patterns of Selection in Cancer and Somatic Tissues. Cell 171(5):1029-1041.

Greenman C, Wooster R, Futreal PA, Stratton MR, & Easton DF (2006) Statistical analysis of pathogenicity of somatic mutations in cancer. Genetics 173(4):2187-2198.
}
\author{
Original author is Inigo Martincorena (Wellcome Trust Sanger Institute)

Qingjian Chen Modified it and add Ki/Ks model. e-mail: \email{chenqingjian2010@163.com}
}
