% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_formate.R
\name{out_formate}
\alias{out_formate}
\title{output formate}
\usage{
out_formate(
  MAF,
  columns = c("Chromosome", "Start_Position", "Reference_Allele", "Tumor_Seq_Allele2",
    "Tumor_Sample_Barcode"),
  mut5 = T,
  vcf = NULL,
  mut5.file = NULL,
  cancer = "cancer",
  columns_Cancer = "Cancer"
)
}
\arguments{
\item{MAF}{MAF files}

\item{columns}{5 column names for output(chr, pos, ref, mut, sampleID).}

\item{mut5}{3) T or F}

\item{vcf}{1) vcf file names}

\item{mut5.file}{2) file names}

\item{cancer}{A simple variable, output cancer name in vcf output.}

\item{columns_Cancer}{Column names in MAF for cancer type. The \emph{columns_Cancer} has \strong{higher priority} than \emph{cancer}.}
}
\value{
mut5 and mut5 file are for dndscv. and vcf is for SSB_selection.
}
\description{
Input is MAF, output is \tabular{lcr}{\cr}
1) vcf; files, vcf  \tabular{lcr}{\cr}
2) mut5.file; Files, simple 5 column, "chr","pos","ref","mut","sampleID", separated by \emph{"Tab"}, without header. \tabular{lcr}{\cr}
3) mut5; system variables, simple 5 column. with header is "sampleID chr pos ref mut \tabular{lcr}{\cr}
}
\details{
\strong{cancer} and \strong{columns_Cancer}: Set Cancer Names in VCF. The Names could be from MAF (\strong{columns_Cancer}) or params  \strong{cancer}
}
