% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removing_patients.R
\name{removing_patients}
\alias{removing_patients}
\title{removing patients}
\usage{
removing_patients(
  MAF,
  patients_col_name = "Tumor_Sample_Barcode",
  maxNum = 3000,
  minNum = 0,
  patientsID = NULL
)
}
\arguments{
\item{MAF}{MAF files}

\item{patients_col_name}{the column name of patients IDs, default is /strong{Tumor_Sample_Barcode}.}

\item{maxNum}{patients with mutations more than this number will be removed.}

\item{minNum}{patients with mutations less than this number will be removed.}

\item{patientsID}{giving the patients IDs and remaining mutations in these patients.}
}
\value{
MAF after removing patients belongs to hyper-mutations
}
\description{
removing patients belongs to hyper-mutations. 
#please keep only SNP mutations. This formulate just count the number, regardless of mutation type.
}
