% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welcome.R
\name{Welcome}
\alias{Welcome}
\title{Welcome}
\usage{
Welcome(name = "")
}
\arguments{
\item{name}{user names}
}
\value{
no return information.
}
\description{
This is my First R Pakcges, The Input is your names.
}
\details{
This is my first R Package
}
\author{
Qingjian Chen (Sun Yat-sen University)
}
